# perf(easy): Hand-optimize `easy::Value` serde support