By default, serde checks each untagged enum's variants, to see if it works.
We can use the serde visitor to directly map types to variants.

This takes decode time for `easy::Value` from 484us to 181us, faster
than decoding directly to a native type (193us) but still slower than
native toml_edit (147us) and toml-rs (121us).